debug models.py

+++ irl/models.py
@@ -44,15 +44,15 @@ class ContextImitation(pl.LightningModule):
     def training_step(self, batch, batch_idx):
         dem_states, dem_actions, test_states, test_actions = batch
         print(dem_actions, dem_states)
-        dem_states.double()
-        dem_actions.double()
-        test_actions.double()
-        test_states.double()
+        dem_states.float()
+        dem_actions.float()
+        test_actions.float()
+        test_states.float()
 
         # concatenate states and actions to get expert trajectory
         print(dem_actions, dem_states)
         dem_traj = torch.cat([dem_states, dem_actions], dim=2)
-        dem_traj.double()
+        dem_traj.float()
         print(dem_traj)
 
         # embed expert trajectory to get a context embedding batch x samples x dim
@@ -97,15 +97,15 @@ class ContextImitation(pl.LightningModule):
     def validation_step(self, batch, batch_idx):
         dem_states, dem_actions, test_states, test_actions = batch
         print(dem_actions, dem_states)
-        dem_states.double()
-        dem_actions.double()
-        test_actions.double()
-        test_states.double()
+        dem_states.float()
+        dem_actions.float()
+        test_actions.float()
+        test_states.float()
 
         # concatenate states and actions to get expert trajectory
         print(dem_actions, dem_states)
         dem_traj = torch.cat([dem_states, dem_actions], dim=2)
-        dem_traj.double()
+        dem_traj.float()
         print(dem_traj)
         # embed expert trajectory to get a context embedding batch x samples x dim
         context_mean_samples = self.context_enc_mean(dem_traj)
